+++ Diary/Scene/DiaryTableViewController.swift
@@ -7,21 +7,11 @@
 import UIKit
 
 final class DiaryTableViewController: UITableViewController {
-    private typealias DataSource = UITableViewDiffableDataSource<Int, Diary>
-    private typealias Snapshot = NSDiffableDataSourceSnapshot<Int, Diary>
-    private var dataSource: DataSource?
-    
-    private var diarys = [Diary]() {
-        didSet {
-            makeSnapshot()
-        }
-    }
+    private var dataSource: DiaryTableViewDataSource?
     
     override func viewDidLoad() {
         super.viewDidLoad()
         setUp()
-        makeDataSource()
-        makeSampleDiarys()
     }
     
     private func setUp() {
@@ -48,10 +38,11 @@ final class DiaryTableViewController: UITableViewController {
         tableView.separatorInset.left = 20
         tableView.register(DiaryCell.self, forCellReuseIdentifier: DiaryCell.reuseIdentifier)
         makeDataSource()
+        dataSource?.makeSampleDiarys()
     }
     
     private func makeDataSource() {
-        dataSource = DataSource(tableView: tableView) { tableView, indexPath, item in
+        dataSource = DiaryTableViewDataSource(tableView: tableView) { tableView, indexPath, item in
             let cell = tableView.dequeueReusableCell(
                 withIdentifier: DiaryCell.reuseIdentifier,
                 for: indexPath
@@ -61,32 +52,13 @@ final class DiaryTableViewController: UITableViewController {
             return cell
         }
     }
-    
-    private func makeSnapshot() {
-        var snapshot = Snapshot()
-        snapshot.appendSections([0])
-        snapshot.appendItems(diarys)
-        
-        dataSource?.apply(snapshot)
-    }
-    
-    private func makeSampleDiarys() {
-        guard let url = Bundle.main.url(forResource: "sample", withExtension: "json") else { return }
-        
-        do {
-            let sampleData = try Data(contentsOf: url)
-            diarys = try JSONDecoder().decode([Diary].self, from: sampleData)
-        } catch {
-            print(error)
-        }
-    }
 }
 
 // MARK: TableViewDelegate
 
 extension DiaryTableViewController {
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let diary = diarys[indexPath.row]
+        let diary = dataSource?.diarys[indexPath.row]
         let diaryViewController = DiaryDetailViewController(diary: diary)
         
         navigationController?.pushViewController(diaryViewController, animated: true)
